sbg-marking-server-py/pyproject.toml
[project]
name = "sbg-marking-server-py"
version = "0.1.0"
description = "Python test stub for the SBG marking server (core endpoints for plugin E2E)."
requires-python = ">=3.11"

dependencies = [
    "Flask==3.0.3",
    "SQLAlchemy==2.0.35",
    "alembic==1.13.3",
    "psycopg[binary]==3.2.3",
    "pydantic==2.9.2",
    "python-dotenv==1.0.1",
    "gunicorn==23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-flask==1.3.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
